feat(redis cache): 缓存序列化设置，添加redis缓存

+++ back-end/Dto/src/main/java/com/xcphoenix/dto/mapper/RestaurantMapper.java
@@ -57,7 +57,7 @@ public interface RestaurantMapper {
      * @param shopId 店铺id
      * @return 店铺信息
      */
-    Restaurant getShopDetailById(Integer shopId);
+    Restaurant getShopDetailById(Long shopId);
 
     /**
      * 店铺是否存在
